recommendations/requirements.txt
Flask
pyjwt
requests